+++ src/app/profile/[slug]/page.tsx
@@ -1,0 +1,9 @@
+import PageProfile from "./PageProfile";
+
+export default function Page({ params }: { params: { slug: string } }) {
+    
+    // by placing in the code the format that mateo wants
+
+    const { slug } = params;
+    return <PageProfile slug={slug} />;
+}
